[Refactor] Simplify paged attention contracts
